testing/web-platform/tests/workers/semantics/interface-objects/001.html












<!doctype html>
<title>available interface objects in dedicated worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, '', 'these interface objects were missing');
    this.done();
  });
});
</script>
